PipeSubclasses.py library
